% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTPnet.R
\name{cTPnet}
\alias{cTPnet}
\title{Imputation of 12 surface proteins from scRNA-seq data.}
\usage{
cTPnet(data, data_type = "Seurat2", model_file_path)
}
\arguments{
\item{data}{data file: can be Seurat object, a matrix or a dataframe.}

\item{data_type}{specify type of data.Need to be one of the four
options: Seurat2, Seurat3, matrix, dataframe. You can check your
Seurat version by sessionInfo()'}

\item{model_file_path}{str: indicate the path to the trained cTPnet pytorch
model for the prediction}
}
\value{
If input is Seurat object, the returned value will also be a Seurat
object; if input is matrix or dataframe, the returned value will be a
dataframe with rows as proteins and columns as cells.
}
\description{
Imputation of 12 surface proteins from scRNA-seq data.
}
