% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shared_gene}
\alias{shared_gene}
\title{Prices of 50,000 round cut diamonds.}
\format{A vector with 12611 genes:
\describe{
  \item{shared_gene}{list of gene names}
}}
\usage{
shared_gene
}
\description{
A data set contain shared gene list which used to train the model.
}
\keyword{datasets}
